Centralise code related to log in into `log_in` module
