+++ src/auth/mod.rs
@@ -1,5 +1,4 @@
 /*! This module defines and implements the data structures, response handlers and functions for authenticating a user and handling cookie auth. */
 
 pub mod cookie;
-pub mod log_in;
 pub mod middleware;
