+++ src/routes/views/mod.rs
@@ -3,7 +3,6 @@
 pub mod dashboard;
 pub mod forgot_password;
 pub mod import;
-pub mod log_in;
 pub mod new_category;
 pub mod new_transaction;
 pub mod register;
